+++ src/business/tdo/create-busin-first.dto.ts
@@ -1,0 +1,25 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsNotEmpty, IsNumber, IsOptional, IsString } from 'class-validator';
+
+export class CreateBusinessDto {
+    @IsString()
+    id: string;
+
+    @IsNotEmpty()
+    @IsString()
+    name: string;
+
+    @IsNotEmpty()
+    owner: Object;
+
+    @IsNotEmpty()
+    @IsString()
+    businessSize: string;
+
+    @IsNotEmpty()
+    @IsString()
+    industryType: string;
+
+    @IsString()
+    email: string;   
+}
